Patched login, fixed build apk issues

+++ lib/backend/database_engine.dart
@@ -27,7 +27,7 @@ class DataBaseEngine {
 
       getLength(firestore);
 
-      firestore.collection('users').getDocuments().asStream().listen((data) {
+      firestore.collection('users').getDocuments().then((data) {
         for (var doc in data.documents) {
           String dataName = doc['name'];
           String dataPassword = doc['password'];
@@ -41,6 +41,8 @@ class DataBaseEngine {
           }
         }
       });
+
+      // firestore.collection('users').getDocuments().asStream().listen((data) {});
     } catch (_) {
       popUpInfo(context, "Error: Connection failed",
           "Unable to connect to the database. Please check your data connection and try again.");
@@ -48,14 +50,12 @@ class DataBaseEngine {
   }
 
   void getLength(var firestore) async {
-    try{
-    var futureAmount =
-        await firestore.collection('users').getDocuments().then((data) {
-      amount = data.documents.length;
-    });
-    }catch(_){
-
-    }
+    try {
+      var futureAmount =
+          await firestore.collection('users').getDocuments().then((data) {
+        amount = data.documents.length;
+      });
+    } catch (_) {}
   }
 
   String createHash(String value) {
@@ -71,24 +71,29 @@ class DataBaseEngine {
       String dataName, String dataPassword, String role, int index) async {
     if ((name == dataName)) {
       try {
-        assert(dBCrypt.checkpw(password, dataPassword), true);
-        if (role == 'Administrator') {
-          isAdmin = true;
-        }
+        if (dBCrypt.checkpw(password, dataPassword)) {
+          if (role == 'Administrator') {
+            isAdmin = true;
+          }
 
-        checkComplete = true;
-        userName = name;
-        roleStatus = role;
+          checkComplete = true;
+          userName = name;
+          roleStatus = role;
 
-        await storage.deleteAll();
-        if (rememberMe == "yes") {
-          await storage.write(key: "name", value: name);
-          await storage.write(key: "role", value: role);
-          await storage.write(key: "rememberMe", value: "yes");
+          await storage.deleteAll();
+          if (rememberMe == "yes") {
+            await storage.write(key: "name", value: name);
+            await storage.write(key: "role", value: role);
+            await storage.write(key: "rememberMe", value: "yes");
+          }
+          Navigator.of(context).pushReplacement(
+              new MaterialPageRoute(builder: (context) => HomeScreen()));
+        } else {
+          throw Exception("Password doesn't match");
         }
-        Navigator.of(context).pushReplacement(
-            new MaterialPageRoute(builder: (context) => HomeScreen()));
       } catch (_) {
+        popUpInfo(context, 'Login Failed',
+            'Incorrect name/password. Please try again.');
         //This is to silence the warning after assert failure
       }
     } else {
@@ -105,8 +110,7 @@ class DataBaseEngine {
     try {
       //CHECK IF DATA EXISTS IN DATABASEE
       firestore.collection(collection).document().setData(data);
-    } catch (_) {
-    }
+    } catch (_) {}
   }
 
   bool validateData(Map<String, String> data) {
